+++ frontend/packages/data-portal/app/components/KeyPhoto.tsx
@@ -1,7 +1,9 @@
 import Skeleton from '@mui/material/Skeleton'
+import { useLocation } from '@remix-run/react'
 import { match, P } from 'ts-pattern'
 
 import { KeyPhotoFallbackIcon } from 'app/components/icons'
+import { useI18n } from 'app/hooks/useI18n'
 import { cns } from 'app/utils/cns'
 
 export function KeyPhoto({
@@ -9,12 +11,17 @@ export function KeyPhoto({
   loading = false,
   src,
   title,
+  overlayOnGroupHover,
 }: {
   className?: string
   loading?: boolean
   src?: string
   title: string
+  overlayOnGroupHover?: boolean
 }) {
+  const { t } = useI18n()
+  const { pathname } = useLocation()
+
   return (
     <div
       className={cns(
@@ -24,8 +31,25 @@ export function KeyPhoto({
 
         // crop image to container dimensions
         'overflow-hidden object-cover',
-
+        overlayOnGroupHover && [
+          'relative',
+          'before:absolute',
+          'before:bg-black',
+          'group-hover:before:bg-opacity-30',
+          'group-hover:before:w-full',
+          'group-hover:before:h-full',
+          'group-hover:before:text-sds-body-s',
+          'group-hover:before:flex-wrap',
+          'group-hover:before:content-center',
+          'group-hover:before:content-i18n',
+          'group-hover:before:text-sds-gray-white',
+          'group-hover:before:font-semibold',
+          'group-hover:before:text-center',
+        ],
         className,
+      )}
+      data-i18n-content={t(
+        pathname.includes('/browse-data') ? 'openDataset' : 'openRun',
       )}
     >
       {match([src, loading])
@@ -38,7 +62,7 @@ export function KeyPhoto({
           />
         ))
         .otherwise(() => (
-          <KeyPhotoFallbackIcon className="text-sds-gray-200 aspect-square w-1/5" />
+          <KeyPhotoFallbackIcon className="text-sds-gray-200 font-normal aspect-square w-1/5" />
         ))}
     </div>
   )