+++ frontend/packages/data-portal/app/components/Link/Link.tsx
@@ -1,10 +1,22 @@
 import { Link as RemixLink, LinkProps } from '@remix-run/react'
 import { ForwardedRef, forwardRef } from 'react'
 
+import { cnsNoMerge } from 'app/utils/cns'
 import { isExternalUrl } from 'app/utils/url'
 
+export const DASHED_BORDERED_CLASSES =
+  'border-b border-dashed hover:border-solid border-current'
+
+export const DASHED_UNDERLINED_CLASSES =
+  'underline underline-offset-[3px] decoration-dashed hover:decoration-solid'
+
 function BaseLink(
-  { to, ...props }: LinkProps,
+  {
+    to,
+    variant,
+    className,
+    ...props
+  }: LinkProps & { variant?: 'dashed-bordered' | 'dashed-underlined' },
   ref: ForwardedRef<HTMLAnchorElement>,
 ) {
   let newTabProps: Partial<LinkProps> = {}
@@ -18,7 +30,19 @@ function BaseLink(
     }
   }
 
-  return <RemixLink ref={ref} to={to} {...props} {...newTabProps} />
+  return (
+    <RemixLink
+      ref={ref}
+      to={to}
+      className={cnsNoMerge(
+        variant === 'dashed-bordered' && DASHED_BORDERED_CLASSES,
+        variant === 'dashed-underlined' && DASHED_UNDERLINED_CLASSES,
+        className,
+      )}
+      {...props}
+      {...newTabProps}
+    />
+  )
 }
 
 export const Link = forwardRef(BaseLink)