+++ frontend/packages/data-portal/app/constants/testIds.ts
@@ -1,4 +1,5 @@
 export enum TestIds {
   EnvelopeIcon = 'envelope-icon',
   OrcIdIcon = 'orc-id-icon',
+  Pagination = 'pagination',
 }