+++ frontend/CHANGELOG.md
@@ -1,5 +1,85 @@
 # Changelog
 
+## [1.10.3](https://github.com/chanzuckerberg/cryoet-data-portal/compare/web-v1.10.2...web-v1.10.3) (2024-06-12)
+
+
+### 🐞 Bug Fixes
+
+* open run on hover ([#799](https://github.com/chanzuckerberg/cryoet-data-portal/issues/799)) ([77a6815](https://github.com/chanzuckerberg/cryoet-data-portal/commit/77a6815bd754618dbbb6e91aa6448d6f37bad3cb))
+
+
+### 🧹 Miscellaneous Chores
+
+* download dialog code snippet ([#794](https://github.com/chanzuckerberg/cryoet-data-portal/issues/794)) ([b88ed71](https://github.com/chanzuckerberg/cryoet-data-portal/commit/b88ed71af37ddd94b715d60750fc8c8de41a48b0))
+
+## [1.10.2](https://github.com/chanzuckerberg/cryoet-data-portal/compare/web-v1.10.1...web-v1.10.2) (2024-06-11)
+
+
+### 🧹 Miscellaneous Chores
+
+* hook up method type to front-end ([#797](https://github.com/chanzuckerberg/cryoet-data-portal/issues/797)) ([d6dcda5](https://github.com/chanzuckerberg/cryoet-data-portal/commit/d6dcda5fd7799d1927efef34c0a6f5ed0d323365))
+* update vite ([#796](https://github.com/chanzuckerberg/cryoet-data-portal/issues/796)) ([7c3e24f](https://github.com/chanzuckerberg/cryoet-data-portal/commit/7c3e24f6d4ea9bc9f991656f10bc8ddacc8d5b1a))
+
+## [1.10.1](https://github.com/chanzuckerberg/cryoet-data-portal/compare/web-v1.10.0...web-v1.10.1) (2024-06-08)
+
+
+### 🧹 Miscellaneous Chores
+
+* remove all mentions of kaggle ([#791](https://github.com/chanzuckerberg/cryoet-data-portal/issues/791)) ([60221d3](https://github.com/chanzuckerberg/cryoet-data-portal/commit/60221d3049a49f799c0b61bd80dc466b6cf3e955))
+
+
+### ⚙ Continuous Integration
+
+* pagination e2e ([#776](https://github.com/chanzuckerberg/cryoet-data-portal/issues/776)) ([3d78ef4](https://github.com/chanzuckerberg/cryoet-data-portal/commit/3d78ef47f6bc5339c69ff2748932740f70ecd1df))
+
+## [1.10.0](https://github.com/chanzuckerberg/cryoet-data-portal/compare/web-v1.9.0...web-v1.10.0) (2024-05-31)
+
+
+### ✨ Features
+
+* add ML Challenge organizers section ([#753](https://github.com/chanzuckerberg/cryoet-data-portal/issues/753)) ([a66f54b](https://github.com/chanzuckerberg/cryoet-data-portal/commit/a66f54b950ed51027b4749eca3392d79938169ff))
+* add ML Challenge Page competition data section and CTA ([#754](https://github.com/chanzuckerberg/cryoet-data-portal/issues/754)) ([e836c22](https://github.com/chanzuckerberg/cryoet-data-portal/commit/e836c226f345d725677338e4f12c6a075c75f98e))
+* add ML Challenge Page FAQ ([#762](https://github.com/chanzuckerberg/cryoet-data-portal/issues/762)) ([30974bf](https://github.com/chanzuckerberg/cryoet-data-portal/commit/30974bf8092585b383a1ab1b63d7757b58cd087f))
+* add ML Challenge page main content ([#739](https://github.com/chanzuckerberg/cryoet-data-portal/issues/739)) ([d5ee77d](https://github.com/chanzuckerberg/cryoet-data-portal/commit/d5ee77d708795ee8deed99fff6067bb1915fff39))
+* ML Challenge Banners ([#755](https://github.com/chanzuckerberg/cryoet-data-portal/issues/755)) ([47e9b45](https://github.com/chanzuckerberg/cryoet-data-portal/commit/47e9b457874cf531b93fbe69bb933537760efd66))
+* ML Challenge Navigation ([#758](https://github.com/chanzuckerberg/cryoet-data-portal/issues/758)) ([131bf6b](https://github.com/chanzuckerberg/cryoet-data-portal/commit/131bf6b278a01d4b0afa49f7e3334c6972bbc234))
+
+## [1.9.0](https://github.com/chanzuckerberg/cryoet-data-portal/compare/web-v1.8.0...web-v1.9.0) (2024-05-30)
+
+
+### ✨ Features
+
+* ux navigation styling updates ([#737](https://github.com/chanzuckerberg/cryoet-data-portal/issues/737)) ([508ab9b](https://github.com/chanzuckerberg/cryoet-data-portal/commit/508ab9b2b1e186c65ab3d0bbf61179eda85139cb))
+
+
+### 🧹 Miscellaneous Chores
+
+* update babel traverse ([#767](https://github.com/chanzuckerberg/cryoet-data-portal/issues/767)) ([f53a779](https://github.com/chanzuckerberg/cryoet-data-portal/commit/f53a7791b108f7ab696e687e19e2b6a47fa74c7a))
+
+## [1.8.0](https://github.com/chanzuckerberg/cryoet-data-portal/compare/web-v1.7.0...web-v1.8.0) (2024-05-28)
+
+
+### ✨ Features
+
+* author orc id link ([#719](https://github.com/chanzuckerberg/cryoet-data-portal/issues/719)) ([9488f2a](https://github.com/chanzuckerberg/cryoet-data-portal/commit/9488f2a126a7bffe63151fa527a77d6df39d8793))
+* ML challenge page routing ([#728](https://github.com/chanzuckerberg/cryoet-data-portal/issues/728)) ([be5e86c](https://github.com/chanzuckerberg/cryoet-data-portal/commit/be5e86c590a139282366ce2b1a06ac0b28fad323))
+* run filter section titles ([#730](https://github.com/chanzuckerberg/cryoet-data-portal/issues/730)) ([5c29681](https://github.com/chanzuckerberg/cryoet-data-portal/commit/5c296814b1fa53bf45b2f30177410119e4e7c092))
+
+
+### 🐞 Bug Fixes
+
+* run header full width for firefox ([#731](https://github.com/chanzuckerberg/cryoet-data-portal/issues/731)) ([28a0738](https://github.com/chanzuckerberg/cryoet-data-portal/commit/28a0738f232be9265db7edb5a1976139cbe53473))
+
+
+### 🧹 Miscellaneous Chores
+
+* primary_annotator_status -&gt; primary_author_status ([#727](https://github.com/chanzuckerberg/cryoet-data-portal/issues/727)) ([a7d5af9](https://github.com/chanzuckerberg/cryoet-data-portal/commit/a7d5af9d6a264daedd43b728fdf68accc3ad9557))
+
+
+### ⚙ Continuous Integration
+
+* e2e filter utils ([#751](https://github.com/chanzuckerberg/cryoet-data-portal/issues/751)) ([21a8bf7](https://github.com/chanzuckerberg/cryoet-data-portal/commit/21a8bf7d3bbff014cbf1c7bcb178b0d964f0ad28))
+
 ## [1.7.0](https://github.com/chanzuckerberg/cryoet-data-portal/compare/web-v1.6.3...web-v1.7.0) (2024-05-10)